refactor: :truck: move all tests to tests folder
