+++ src/tests/components/Icon/index.spec.tsx
@@ -1,5 +1,5 @@
 import renderer from 'react-test-renderer'
-import Icon from '.'
+import Icon from '../../../components/Icon'
 
 it('renders correctly', () => {
   const tree = renderer.create(<Icon icon="logo" />).toJSON()
