+++ src/tests/components/Layout/index.spec.tsx
@@ -1,5 +1,5 @@
 import renderer from 'react-test-renderer'
-import Layout from '.'
+import Layout from '../../../components/Layout'
 
 it('renders correctly', () => {
   const tree = renderer.create(<Layout path="home" />).toJSON()
