+++ src/tests/pages/technology/index.spec.tsx
@@ -1,5 +1,5 @@
 import renderer from 'react-test-renderer'
-import Technology from '.'
+import Technology from '../../../pages/technology'
 
 it('renders correctly', () => {
   const tree = renderer.create(<Technology />).toJSON()
